Removing article link "This topic is no longer available"
<div>Read the full article on <a title="MSDN" href="http://msdn.microsoft.com/en-us/library/hh532261(v=VS.103).aspx" target="_blank"> MSDN</a>.</div>

*This link is broken:* http://msdn.microsoft.com/en-us/library/hh532261(v=VS.103).aspx
Not sure of the correct link to replace it with, and/or if instead it should be omitted altogether as it may no longer be relevant?

Also as I understand it the Windows Azure Management Portal is now referred to either as the "classic" portal: https://manage.windowsazure.com or the "new" portal: https://portal.azure.com/ but didn't want to change any URLs. Does the Service Bus Explorer interacts in different ways depending on which portal and their different functionalities?