- Replaced TextBox controls used for reading and writing the payload of messages with FastColoredTextBox  - Fixed an error in Service Bus Relay test control  - Changed Azure icon in the title of all forms  - Upgraded WindowsAzure.ServiceBus NuGet package to version 4.1.8  - Upgraded Newtonsoft.Json NuGet package to version 11.0.0
Note: the dll containing the FastColoredTextBox control is a customization of the control available on github at https://github.com/PavelTorgashov/FastColoredTextBox.
For the momement, I just referenced the dll and added the dll in a folder called Library. I know that adding binaries to a project on GitHub is not a good practice,
but I cannot create a NuGet package for the customization of a library that is already available as a NuGet package. Thoughts?